Allow for customization of waking request
In order to ensure that an image displays, we must send a command that
we know will "wake" Kodi up, since Kodi does not always do so when an
image is opened. The default is "Input.Left" which is just a directional
button press and is not likely to cause problems for users.